refactor(bootstrap): extract WiredAdapters + LifecycleResources to bootstrap/types.py
Both types previously lived in factory/ and lifecycle/ subpackages, creating a
real logical cycle (lifecycle imported WiredAdapters from factory via TYPE_CHECKING;
factory/unified imported LifecycleResources from lifecycle at runtime). Moving both
to a neutral bootstrap/types.py eliminates the cross-subpackage dependency entirely.

Closes #997
