+++ src/lyra/bootstrap/lifecycle/bootstrap_lifecycle.py
@@ -5,13 +5,6 @@ from __future__ import annotations
 import asyncio
 import logging
 import os
-from dataclasses import dataclass, field
-from typing import TYPE_CHECKING
-
-if TYPE_CHECKING:
-    from nats.aio.client import Client as NatsClient
-
-    from lyra.bootstrap.factory.wiring_helpers import WiredAdapters
 
 import uvicorn
 
@@ -21,23 +14,10 @@ from lyra.bootstrap.lifecycle.lifecycle_helpers import (
     teardown_buses,
     teardown_dispatchers,
 )
-from lyra.core.cli.cli_pool import CliPool
+from lyra.bootstrap.types import LifecycleResources, WiredAdapters
 from lyra.core.hub import Hub
-from lyra.infrastructure.stores.pairing import PairingManager
-from lyra.nats.nats_channel_proxy import NatsChannelProxy
 
 log = logging.getLogger(__name__)
-
-
-@dataclass
-class LifecycleResources:
-    """Optional infrastructure wired into the lifecycle."""
-
-    pm: PairingManager | None
-    # unified: always None; CliPool managed in unified.py finally
-    cli_pool: CliPool | None
-    proxies: list[NatsChannelProxy] | None = field(default=None)
-    nc: NatsClient | None = field(default=None)
 
 
 async def run_lifecycle(  # noqa: C901 — lifecycle orchestration
